Some fixes and added PCBs
I haven't looked at this in a while but here are some PCBs and a few bug fixes to the code. The code is about to get an overhaul to make it more readable but is working as is.

The PCBs are in use except for the SOIC version which I haven't tested yet.